+++ CHANGELOG.md
@@ -13,7 +13,7 @@ This release contains the following:
 - Support for multiple protocol IDs - now matches any protocol that adds postfix to stable ID.
 - Updates the order of fields of `HistoryResponse` protobuf message. 
   The filed numbers of the `HistoryResponse` are shifted up by one to match up the [13/WAKU2-STORE](https://rfc.vac.dev/spec/13/) specs. 
-
+- Adds optional `timestamp` to `WakuRelayMessage`.
 #### General refactoring
 #### Docs
 - Adds the database migration tutorial.